Fix bug in qr_decomposition, more remains.
The bug fixed was R being non upper-traingular because it was not zero
initalized, and the algorithm only modifies the upper triangle of the
matrix, so trash values were left over. With this fix R in the
decomposition is now correct (as far as my testing shows).

The matrix Q in the qr_decomposition is still incorrect: according to
testing, it's last column exactly differs from the correct Q (from
np.linalg.qr). Despite that error Q@R actually still results in the original matrix, however Q@Q.T differs from the identity matrix meaning Q is not orthogonal as desired.