Fix qr_iteration bug, improve (debugging) code.
The bug fixed was qr_iteration breaking on the wrong condition,
||q|-|q'q||<EPSILON was checked instead of ||q'|-|q'q||<EPSILON.
This solves U in the spectral clustering algorithms having zero rows,
which (almost?) always happened previously.

Also added TODO and easy if True else expression for choosing to add a
small number to the diagonal of the normalized graph Laplacian matrix
before using qr iteration on it - this can better the accuracy of the
qr_decomposition algorithm, whose last column differed from the numpy
function because of an inherent algorithm flaw (according to the forum).

According to testing there's a good chance with this we've gotten
rid of all the bugs in the new algorithms (everything introduced in the
final project - the contents of the spectral clustering algorithm, not
k-means).

Debugging code was also improved, the code is pretty self-explanatory.